Fix erroneous requirement of sample_rate field in Channel
The specification does not require the presence of a
SampleRateGroup, but if one is present, then there must be
a SampleRate element and may be a SampleRateRatio.

Make the sample_rate field a Union{Missing,SampleRate},
but do not enforce the presence of sample_rate if sample_rate_ratio
is given.