Add `warn` option to `write`
When writing StationXML files, pass `warn=true` as a keyword
argument to emit warnings.  At present, warnings are emitted
when a `FDSNStationXML` object is written, but has been
read from a file with an older version of the StationXML
specification, and when it contains fields (elements or
attributes) which have been removed from the spec.  In this
case, information will be lost on writing.

The default for `warn` is `false`.