Warn instead in ResponseStage when giving polynomial and stage gain
Although v1.1 of the spec disallows the presence of a stage gain in
a polynomial stage, v1.0 does not expressly disallow it, so rather than
throw an error as in commit 05367d0, just emit a warning.

In future, it may be worth writing specific methods for types which
have changes like this which must be enforced between versions
of the specification.