Add merging of and appending to `FDSNStationXML` objects
Add methods to `Base.merge!`, `Base.merge` and `Base.append!` to allow
multiple `FDSNStationXML` objects to be merged or concatenated together.

'Merging' using `merge[!]` will attempt to remove duplicate networks, stations
and channels during the process, so that what remains is only the unique,
non-overlapping content of the two objects.  Currently if there is an overlap
between channels at the same station (determined by station code only), then
we warn the user and do not add the channels with duplicate informaiton.

As an example, if two files contain channels:
  1. AA.BB..CC
  2. AA.BB..DD
then the resulting object will have a single network 'AA', a single station
'BB' and a two channels within that one station, 'CC' and 'DD'.

The algorithm to determine how to merge is subject to change in the
future!

As an alternative, the user can also `append!` two sets of metadata
together.  This simply appends the networks of one set to the other.  If there
are duplicated networks, stations or channels, they end up repeated.
This is the behaviour implemented by ObsPy.