Fix error in overloading getproperty(AbstractArray, ...)
In the package, we add methods to `Base.getproperty` for
`AbstractArray`s of types we own in StationXML.jl.  Up until now
this hasn't caused any problems, even if it may be type piracy.

However as of Julia v1.11, `Array` is now a Julia type and as such
`getproperty(::Array, ::Symbol)` is defined, which breaks the code
here.  See https://github.com/JuliaLang/julia/issues/56100 for
details.

Update the existing `getproperty(::AbstractArray, ::Symbol)`
methods and add new ones for `Array` to overcome this problem.

These methods will be removed in future to avoid type-piracy.